requirements.txt
amqp==2.5.1
billiard==3.6.1.0
blinker==1.4
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-paginate==0.5.3
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.4
MarkupSafe==1.1.1
numpy==1.17.0
pandas==0.25.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
sentry-sdk==0.10.2
six==1.12.0
SQLAlchemy==1.3.7
urllib3==1.25.3
vine==1.3.0
Werkzeug==0.15.5
zipp==0.5.2
